+++ SpringBootCodexBoxJava2022Batch1/src/main/java/com/codexbox/springboot/app/sohailspringbootproject/basicexamples/Factorial.java
@@ -1,6 +1,19 @@
 package com.codexbox.springboot.app.sohailspringbootproject.basicexamples;
 
-public class Factorial {
-    int fact=0;
+import java.util.Scanner;
 
-}
+public class Factorial {
+    public void factorialMethod(){
+        int x= 0;
+        Scanner scanner = new Scanner(System.in);
+        System.out.println("Enter the number of factorial you want : ");
+        x=scanner.nextInt();
+        for(int i = 0 ; i<x;i++) {
+
+        }
+            System.out.println();
+        }
+
+    }
+
+
