--- SpringBootCodexBoxJava2022Batch1/src/main/java/com/codexbox/springboot/app/sohailspringbootproject/examples/Factorial.java
@@ -1,19 +1,0 @@
-package com.codexbox.springboot.app.sohailspringbootproject.examples;
-
-import java.util.Scanner;
-
-public class Factorial {
-    public void factorialMethod(){
-        int x= 0;
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("Enter the number of factorial you want : ");
-        x=scanner.nextInt();
-
-
-
-            System.out.println();
-        }
-
-    }
-
-
