+++ SpringBootCodexBoxJava2022Batch1/src/main/java/com/codexbox/springboot/app/sohailspringbootproject/main/MainMethod.java
@@ -1,6 +1,6 @@
 package com.codexbox.springboot.app.sohailspringbootproject.main;
 
-import com.codexbox.springboot.app.sohailspringbootproject.examples.Factorial;
+import com.codexbox.springboot.app.sohailspringbootproject.basicexamples.Factorial;
 
 public class MainMethod {
     public static void main(String [] args ){
